Reworking Our Work Page
Got thumbnails to show up and scrapped the youtube api design for security reasons. The page will in the future have thumbnails with descriptions out to the side. Additionally, when you hover the thumbnail will gray out with a youtube icon over the thumbnail. When clicked it would take the user to the actual video. All things yet to come.