Home Page Revamped
The home page for the website has been made better. There is a placeholder for the big touch studios logo in the middle of the navbar. Additionally, there is a collection of screenshots from the various works listed in the Our Work page. The home page needs some work with the slideshow as when the final hits it doesnt fade in. Additionally I want it to all fit and not need the user to scroll down to see everything.