Our Work Change + Bug Fix
Changed the Our Work page so that the white space is less noticeable. Text is now under the video with the video itself becoming bigger. There needs to be more design changes but this is good for now. Additionally, the issue with index before was just from how github pages works per setup. Will fix later.